requirements/common.txt
matplotlib==3.10.8
numpy==2.4.0
Pillow==12.0.0
psutil==7.1.1
scikit-learn==1.8.0
seaborn==0.13.2
tqdm==4.64.1
